backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
psycopg[binary]==3.2.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
redis==5.1.1
httpx==0.27.2
scikit-learn==1.5.2
spacy==3.7.6
google-generativeai==0.8.3
python-jose[cryptography]==3.3.0
